requirements.txt
fastapi==0.116.1
jose==1.0.0
motor==3.7.1
pydantic==2.11.7
pymongo==4.13.2
python-dotenv==1.1.1
python_bcrypt==0.3.2
python_jose==3.4.0
uvicorn==0.35.0
